1. 一种没有上下文切换的快速ipc通信机制原型，非常快！！！   2. 增加fork_exec系统调用，将fork和exec合并，在单片机上非常适用，能够节约内存。

+++ mkrtos-real/lib/dietlibc/include/sys/types.h
@@ -56,10 +56,10 @@ typedef unsigned long int clock_t;	/* Used for system times in
 //    typedef uint16_t nlink_t;
 //    typedef uint16_t uid_t;zhangzheng
     typedef uint32_t dev_t;
-    typedef uint32_t gid_t;
+    typedef uint16_t gid_t;
     typedef uint32_t mode_t;
-    typedef uint32_t nlink_t;
-    typedef uint32_t uid_t;
+    typedef uint16_t nlink_t;
+    typedef uint16_t uid_t;
 #elif defined(__hppa__)
     typedef uint32_t dev_t;
     typedef uint32_t gid_t;
